Refactor input and output configuration components
